+++ Libplanet/Blocks/Block.cs
@@ -532,6 +532,32 @@ namespace Libplanet.Blocks
                 : OrderTxsForEvaluationV0(txs, preEvaluationHash);
         }
 
+        // FIXME: This should be moved to BlockContent<T>.TxHash.
+        internal static HashDigest<SHA256>? CalculateTxHashes(IEnumerable<Transaction<T>> txs)
+        {
+            if (!txs.Any())
+            {
+                return null;
+            }
+
+            byte[][] serializedTxs = txs.Select(tx => tx.Serialize(true)).ToArray();
+            int txHashSourceLength = serializedTxs.Select(b => b.Length).Sum() + 2;
+            var txHashSource = new byte[txHashSourceLength];
+
+            // Bencodex lists look like: l...e
+            txHashSource[0] = 0x6c;
+            txHashSource[txHashSourceLength - 1] = 0x65;
+            int offset = 1;
+            foreach (byte[] serializedTx in serializedTxs)
+            {
+                serializedTx.CopyTo(txHashSource, offset);
+                offset += serializedTx.Length;
+            }
+
+            using SHA256 hashAlgo = SHA256.Create();
+            return HashDigest<SHA256>.DeriveFrom(txHashSource);
+        }
+
         /// <summary>
         /// Validates this <see cref="Block{T}"/> and throws an appropriate exception
         /// if not valid.
@@ -584,31 +610,6 @@ namespace Libplanet.Blocks
                 header: Header,
                 transactions: Transactions
                 .Select(tx => tx.Serialize(true).ToImmutableArray()).ToImmutableArray());
-        }
-
-        private static HashDigest<SHA256>? CalculateTxHashes(IEnumerable<Transaction<T>> txs)
-        {
-            if (!txs.Any())
-            {
-                return null;
-            }
-
-            byte[][] serializedTxs = txs.Select(tx => tx.Serialize(true)).ToArray();
-            int txHashSourceLength = serializedTxs.Select(b => b.Length).Sum() + 2;
-            var txHashSource = new byte[txHashSourceLength];
-
-            // Bencodex lists look like: l...e
-            txHashSource[0] = 0x6c;
-            txHashSource[txHashSourceLength - 1] = 0x65;
-            int offset = 1;
-            foreach (byte[] serializedTx in serializedTxs)
-            {
-                serializedTx.CopyTo(txHashSource, offset);
-                offset += serializedTx.Length;
-            }
-
-            using SHA256 hashAlgo = SHA256.Create();
-            return HashDigest<SHA256>.DeriveFrom(txHashSource);
         }
 
         private static IEnumerable<Transaction<T>> OrderTxsForEvaluationV0(