+++ CHANGES.md
@@ -17,12 +17,15 @@ To be released.
 
 ### Added APIs
 
+ -  Added `BlockContent<T>` class.  [[#1164], [#1492]]
+
 ### Behavioral changes
 
 ### Bug fixes
 
 ### CLI tools
 
+[#1164]: https://github.com/planetarium/libplanet/issues/1164
 [#1492]: https://github.com/planetarium/libplanet/pull/1492